requirements.txt
azure-core==1.30.1
azure-functions==1.19.0
azure-identity==1.16.0
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
portalocker==2.8.2
pyarrow==16.0.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
